Android: fix Dialog window decorations' dismiss-on-click behavior
This fixes the case when a modal dialog is instantiated in a visible
state, which often means, the window decoration component has not been
loaded yet.

Previously, this led to dismissOnOuterClick property in
DefaultWindowDecoration keeping its initial value of "true" and thus
allowing dismissing the dialog via an outside click *the first time
the dialog was hidden* (next times, the property would get updated
and the behavior would be correct).

Task-number: QTBUG-47369
Change-Id: I24c7c8a1eb785203a3b3f1d889b7e22480ea952f
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>