Check for __style existence before accessing it
Since __style is loaded by a Loader, very early in initialization, it
may still be null. so check for its existence before accessing it.
When it will be fully loaded the property binding will cause a
reevaluation with the proper values.
This fixes __wheelScrollLines never being loaded from the style.

Change-Id: I5967265a4d31e1be2c972daa79b9389fecc25933
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>