TreeView should update selection on release with ExtendedSelection
The selection shouldn't change on press if multiple items are selected and
the user presses on an already selected item. This is the correct behavior
on Windows and OSX where selection changes on release.
Furthermore reacting on press prevents the end user
to implement dragging. In fact if we react on press
the selection changes before the dragging event starts.

Change-Id: Iff5bca8efa5f0701c479f63ff8b1f442268a98d5
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>