Port to new Q*DialogOptions API
Q*DialogOptions are no longer value types, but are
held only in QSharedPointers. Consequently, all
special member function have been hidden from clients
and create() and clone() methods have been added.

This change uses a macro defined by the QtBase change
to select the API variant.

This is needed for the QtBase change to integrate, and
needed for as long as the QtBase change hasn't been
integrated, yet.

A follow-up commit will remove the scaffolding again.

Change-Id: I20c1d75fcaa42323862599f2a35b815bc67b39dd
Reviewed-by: Shawn Rutledge <shawn.rutledge@theqtcompany.com>