ComboBox: guarantee index consistency when emitting activated()
The "triggered" signal from the ComboBox menu items is handled twice:
once in ComboBox.qml, and once in QQuickMenuItem. The function of the
latter is to update the parent QQuickMenu's selected index.
With the current change we guarantee that the selected index (that is
exposed as "currentIndex" in the public API) is updated only once,
before any activated() signal is emitted. This also allows the handler
of the activated() signal to update currentIndex at will, without
running the risk of having it overwritten again by the ComboBox internal
code.

Task-number: QTBUG-51113
Task-number: QTBUG-43050
Change-Id: Ibfe12752d02eb4c9e2bf9a73ee8b3ab0ca8d7eaa
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>