Fix moving of TreeView items
Property binding for row property in styleData causes an update which
tries to read new value for the index property, but index is changed
afterwards which causes old value to be read. This may lead to crashes
and other unwanted behavior.

Depth changes are now delivered to update item depths in visible
items and model index changes though role instead of looking for a row
change.

Task-number: QTBUG-47523
Change-Id: I540cd06a25281f18e4628f4b030cf969dc8e0a7f
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>