Dialogs: enforce changes in minimum size
On OS X, if visible: true is declared, the dialog can be shown before
the changes in 10f0ae7c571b7e0de41d5fdc5627db6a1ddeb926 are able to
set the minimumWidth/Height properties.  QWindow::setMinimumWidth()
also does not force the window to be wider if it is already shown.

Task-number: QTBUG-49058
Change-Id: Id87d92072904555639e1695851d908550d22fe86
Reviewed-by: Alberto Mardegan <mardy@users.sourceforge.net>
Reviewed-by: Nikita Krupenko <krnekit@gmail.com>
Reviewed-by: Liang Qi <liang.qi@qt.io>