ScrollView: block updates when redoing the layout
Changing the contentHeight might change the maximum value of the
scrollbar which might change the value of the scrollbar which, if
blockUpdates is not set, will change the contentY of the flickable.

This can cause flickering when contentHeight varries with contentY
which might typically happen for a ListView where not all elements
have the same height. Or worse, this can make the scrolling jump
and result in the wrong position.

Change-Id: I1968055492d679387ebbf6813a160efbf82e4fbb
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@qt.io>