macOS: Fix native dangling menu still visible on screen and crash
If the parent window gets destroyed while a QtQuick Controls menu is open
the macOS native platform menu is not dismissed and you see a blank
gray rectangle without any text.
Also, at this point the QQmlEngine was already destroyed but it's still
present on the call stack, so you get a crash when the stack unwinds to
the original right mouse click that created the context menu.

Change-Id: I638b0de13734815995d2994e6dd6603bcb0ebefc
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@qt.io>