Fix ColorDialog.color and FontDialog.font update on accept
When using native dialogs, ColorDialog.color was not updated at all,
and FontDialog.font was incorrectly following FontDialog.currentFont.
Promote the overridden accept() methods to the abstract base classes
so the native dialogs get the appropriate property updates on accept
the same way compared to the non-native dialogs.

Task-number: QTBUG-55298
Change-Id: I9ac8f5ecc60884cd98b58d09ef3dcb4baf47772d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>