TumblerStyle: deprecate spacing property
Not only was this unused, but the idea doesn't make sense for Tumbler,
which is internally based on PathView. Having gaps between delegates in
a PathView means that there will be areas where the user cannot "flick"
the Tumbler.

The height of the tumbler can be increased and/or visibleItemCount can
be decreased in order to achieve larger spacing between text.

[ChangeLog][Styles][TumblerStyle] Deprecated unused spacing property.

Task-number: QTBUG-55289
Change-Id: Ieadaac8678642d747820b201598745c7e099ead4
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@theqtcompany.com>