Slider: fix the handle's position when minimumValue is not 0
__handlePos (badly named; it should be, e.g. __handleValueForPos)
represents the value of the handle based on "fakeHandle"'s position. It
is the result of range.valueForPosition() being called. However, the
arguments to this function (fakeHandle.x and fakeHandle.y) don't always
change when the values of the slider change, which leads to the x
calculation for the handle delegate in SliderStyle having outdated
information, causing the related bug.

The fix for another bug already works around this issue by passing the
relevant properties as arguments (which are ignored) to the function
call. This is presumably done this way because it should be cheaper
than forcing the JavaScript engine to evaluate a more clearly written
expression where each related property is on its own line, for example.

property real __handlePos: {
    range.positionAtMinimum, range.positionAtMaximum;
    return range.valueForPosition(__horizontal ? fakeHandle.x : fakeHandle.y);
}

In the case of the related bug, minimumValue has been updated, but
__handlePos is still using the old value, causing the handle to be
positioned incorrectly. So, we continue this tradition and add another
property to the list of arguments.

Task-number: QTBUG-51765
Change-Id: I40882872e668a867a8f5e5768244e199618bd769
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>