Fixed possible crash due to out of memory on ARM
When diving two real values we can obtain an Infinite value.
The qml engine perform an implicit conversion (and silent) if this value
is directly assigned to an "int" property. On Arm the conversion of
Infinite to int gives the value +2147483648. Thus the Repeater model
instantiate 2147483648 items and crashes. By morphing  both the rowHeight
and paddedRowCount properties to "real" we keep the Infinite value semantic.
This allows the Repeater to handle properly the Infinite case.

Task-number: QTBUG-57283
Change-Id: I376f9ca497bea1f1aab413d1c1ba87d918b73fbb
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>