TreeView: allow selecting items by touch
Only single items can be selected via an actual touchscreen, because
the keyboard modifiers are not accessible in a mouse event synthesized
from touch.  With a mouse, if the system has a touchscreen, then
selections done by consecutive clicking (with or without modifiers)
will work, whereas dragging across a consecutive range will not work
because dragging is the same as flicking.  But on systems which do
not have a touchscreen at all, the behavior is the same as before.

Task-number: QTBUG-47243
Change-Id: Ib0fa1a75592b982fe93da46f0c2e3018219947d0
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@qt.io>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>