Slider: add property to disable scroll wheel
Add a property to disable the scroll wheel for the Slider.
This is useful if there is a scrollable list of sliders, because scrolling
would be interrupted as soon as the mouse gets above a slider
(whereby scrolling would change the slider value).
By setting this property to false, a user can easily scroll through the list.

[ChangeLog][Slider] add property to disable scroll wheel

Task-number: QTBUG-59271
Change-Id: I9306638ef98f8a0a5035709a67cc1ee40713441a
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>