Add NOTIFY signals for WheelArea
the private WheelArea component didn't have any notify signals
in its horizontalValue/verticalValue or minimum/maximum
values as well. this caused old values to be used in ScrollView.qml.
If we want to have a listView that starts scrolled right when instanced,
verticalMaximumValue won't be updated when the list is populated.
Also, the wheelArea.verticalValue that's assigned at the beginning will
be constrained by the wrong verticalMaximumValue, causing it to go out of
sync with contentY. At this point the first time the mouse wheel is
moved, the list will jump back at the top.

Change-Id: I4605000636be7975ba9a58e2c79e8c2351e5a292
Task-number: QTBUG-59633
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>