Slider: Don't suppress updates by dragThreshold on non-touchscreens
When using a mouse, you want a slider to react immediately to input,
not only if you move beyond a certain threshold. With the threshold it
is very hard to accurately position the slider. On touchscreens, the
threshold is necessary to allow for unrelated events, like flicking
views, to not interfere with the value.

Change-Id: I6cb8252338f20de559e062e738118593ffc1e099
Task-number: QTBUG-47081
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>