Fix inclusion of QML files of the Dialogs module in static builds
This is a regression introduce by commit
751f08cf46f505980b8ebe9a228a6c229b470152. For a build with QML
Ahead-of-Time cache creation, we need the qml files in QML_FILES and
qml_module.prf takes care of installing them correctly into the file
system (caches and sources). For a static build we also need the sources
in QML_FILES, in order for their automatic inclusion in the resource
system (also by qml_module.prf).

Task-number: QTBUG-60861
Change-Id: Id9125530d8b385101994600c1224dbcb2a5333f9
Reviewed-by: Andy Shaw <andy.shaw@qt.io>