Pick the default style name at runtime instead of compile time
Instead of a naïve OS-based compile time check, we check which platform
plugin is in use at runtime. This should give a more accurate mapping,
especially on desktop OSes using alternative platform plugins where the
Desktop style is not useful.

The minimum Android SDK version is 16 now, so the check is removed.

Change-Id: Ie7af3a8ce5fd031256f5eba9706f24ab50a23bf9
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@qt.io>