Add internal types to Android styles qmldir
In order for types in a module to be available inside the module
itself, but not to external types, they need to be marked as
internal in the qmldir. This has worked by accident before, but
stopped working after 22a2cc43387ec3b9f74a6c01f8665378a4541147
in qtdeclarative, at which point no Qt Quick Controls applications
would run on Android.

Note that there is also a change needed in qtdeclarative, in order
to support implicit loading of internal types even when there is
an explicit import of the module (needed to get access to the
singleton).

Task-number: QTBUG-63309
Change-Id: I36b9ce9fa8f55c4da1f5e64c8fed120e9913bc85
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>