ScrollView: fix cyclic Tab navigation
The ScrollView is not a control widget, which should be included in the
focus chain when the user presses the Tab key; instead, it's a container
and as such it should not interfere with the navigation.

We also modify TableView, because it derives from ScrollView but we
want to keep its current behavior.

Task-number: QTBUG-64596
Change-Id: Ibd7833603d38171693b2f34c5859e9c4615b8ed4
Reviewed-by: Liang Qi <liang.qi@qt.io>