Fix tst_combobox
There was a bogus timer firing once after the test had been running 500 ms. The
timer was supposed to be used as a workaround on Mac, but since it had running
set to true, it was actually firing once on all platforms.

The timer was originally added in 6254af5. The test function ended up being
skipped on Mac, so the timer was actually pointless.

Task-number: QTBUG-64664
Reviewed-by: Liang Qi <liang.qi@qt.io>
(cherry picked from commit a749525164b771452599e67e277d2ea6b3866479)
Change-Id: Icebda05c872ba2de690e87bd0a9968670f203ce2
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>