Fix tst_combobox
There was a bogus timer firing once after the test had been running 500 ms. The
timer was supposed to be used as a workaround on Mac, but since it had running
set to true, it was actually firing once on all platforms.

The timer was originally added in 6254af5. The test function ended up being
skipped on Mac, so the timer was actually pointless.

Task-number: QTBUG-64664
Change-Id: I5b4803ba49db0ab7e1b4dd632a34d51af59bbf91
Reviewed-by: Liang Qi <liang.qi@qt.io>