TableView: Don't unset the delegate loader source component
Once we've loaded the delegate, we want to keep it. Constantly deleting
and recreating items is rather wasteful.

The items could already be moved to a different location before, so
keeping them around when the item is completely disabled doesn't make
much of a difference. We do retain the behavior of only instantiating
the delegate once we add it to the table, though.

Change-Id: Ic3cd016bc6990f0a9695b5aeb0d1de3aad0f4ca1
Task-number: QTBUG-62809
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>
Reviewed-by: Louis du Verdier <louis.du.verdier@free.fr>