Add creation benchmarks for qmlbench
The qmlbench repo has benchmarks only for a subset of QQC2. They are
part of a larger collection of benchmarks that are run periodically, for
which the results are collected to Grafana. We have limited resources
there, so we don't want to overload it with benchmarks for many QQC1
types. However, for our own benchmarking purposes it is useful to be
able to easily benchmark these types.

Change-Id: I9124791a93f07346dc80f1f08aedd655ed0cfa02
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>