Fix memory leak in QQC1
When repeatedly creating a QQuickView, loading a QML file that imports
QQC1 and deleting the view again, we would leak memory that was
allocated as a consequence of QML type registration in
initializeEngine() callbacks that were called on every iteration.

After the limitation of namespacing in the registerTypes() callback of
QML module plugins has been lifted, we can move the type registrations
into registerTypes() where they belong and which is called only once.

Change-Id: I1d21b18dd060e169a1ee1095f5c5af7341443492
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Michael Winkelmann <michael.winkelmann@qt.io>