Calendar: make clicked(date) emit the correct date
Calendar was implemented in such a way that pressing on a date selects
that date, rather than releasing or clicking. With the current code,
this presents issues when the month changes as a result of the press.
For example, clicking on a date in an adjacent month will result in the
clicked() signal passing the date under the mouse in the new month,
instead of the originally pressed date.

This patches fixes the issue by storing the pressed date and using it
when emitting the released() and clicked() signals.

Task-number: QTBUG-54129
Change-Id: I0c16293033b77f6ae783b5365d198b4a516af90b
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>