extras/gallery example: Use layout alignment in Dial page
Replace 'anchors' by the 'Layout.alignment' attached property,
fixing numerous warnings:
qrc:/qml/gallery.qml:137:17: QML Dial: Detected anchors on an item that is part of a layout. This is undefined behavior.
qrc:/qml/gallery.qml:151:17: QML Text: Detected anchors on an item that is part of a layout. This is undefined behavior.
qrc:/qml/gallery.qml:103:17: QML Dial: Detected anchors on an item that is part of a layout. This is undefined behavior.
qrc:/qml/gallery.qml:124:17: QML Text: Detected anchors on an item that is part of a layout. This is undefined behavior.

Change-Id: Iaf5b8f16e477ecd8e7598e857de0fc16e7b53d53
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>