Dialogs: allow preventing processing of standard buttons
Add an actionChosen() signal to DefaultDialogWrapper, allowing the
developer to stop further processing of the event by setting the
"accepted" field to false. This is especially useful when some
validation on the dialog contents need to take place before the dialog
can be accepted.

[ChangeLog][Dialogs] Add a signal to the Dialog class to allow the
client to intercept the button presses and optionally prevent further
processing of the event. This allows performing some validation on the
fields before dismissing the dialog.

Task-number: QTBUG-69095
Change-Id: I19bca0bd9fcbafc72d337a5870776a96634ba748
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>