Document the default value of TableView's overridden frameVisible property
The property is actually already listed in the inherited members
section of the documentation, but the default value listed there
is false, which is not the case for TableView since it overrides it.

Add documentation for the property in TableView, so that it shows
up in the inherited section as defaulting to false, and again in
TableView's own properties section as defaulting to true.

Task-number: QTBUG-64049
Change-Id: Ia3d304cdf1bd34f9d88af388ae04a3ed4a686fbd
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>