ScrollView: Defer calls to doLayout()
doLayout() updates the dimensions of the scroll view, and therefore can
trigger the scroll bars to be shown or hidden. That, in turn, can
trigger another call to doLayout(), resulting in binding loops. Avoid
those by deferring the call.

The scrollview tests are broken without this change because the binding
loop prevents the scrollbars from being shown or hidden in some cases.
The deferring necessitates a tryVerify(), but fixes the actual tests.

Change-Id: I0dac0ba380240e025237f69c4d357d764a535c3d
Fixes: QTBUG-50605
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>