Menu: Provide a safe clear() function
The menu items are generally owned by QML. Therefore we cannot delete
them on clear(). The containers, however, are owned by the menu, and we
have to delete them also when clearing the QQmlListProperty for a proxy
menu. The QQmlListProperty, when clearing the items for a non-proxy
menu, will still delete them, as that seems to be the way QML expresses
its desire to get rid of them.

Fixes: QTBUG-64464
Change-Id: Ied0b86496b289c2e2a83f3d6fd53d7045929beb0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>