QQuickTreeModelAdaptor1: fix out of range issue in selectionForRowRange
If the index of the "to" row is invalid and the one from the "from" row
is good, then the method will try to access an invalid index. This
patch fixes that by ensuring that if the "to" row is invalid (when, for
example, the mouse cursor goes out of the view while multiple selection
is activated), the "to" index used stays valid.

[ChangeLog][Controls][QQuickTreeModelAdaptor1] Fixed an issue where an
invalid access was done if either extended or multiple selection was
in use and the mouse went out of the view on top of the model.

Fixes: QTBUG-71789
Change-Id: I32d8a20ee9ffc1c40584f17c27ceb06d48b81ff0
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>