Insert before next item when it exists
Something is off about the way the last child of previous item is
calculated, but we can work-around that by just using the next element
when it exist. In the case there isn't a next element, we will hit the
fallback in 'last child of previous' and insert at the end anyway.

Fixes: QTBUG-66062
Change-Id: Iced69d52c4587434ffdbb09b08b3441289f34eba
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>