Calendar: Ignore any time specified for a minimum and maximum date
Since Calendar only cares about the date part and not the time, we
should ignore the time part when it is set as the minimum and maximum
and manually set it to 0:00:00 and 23:59:59 respectively. This ensures
that the day itself is still seen as valid irrespective of what time
was passed to Date().

Change-Id: I57c311ffe7d3fb1ab69c59296a067a66f1007137
Fixes: QTBUG-71997
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>