Static: Install the qmldir for the Styles module
Mention qmldir in AUX_QML_FILES to ensure that it is always
copied to the build directory. qml_module.prf usually takes care of this
by having QML_FILES in qmldir.files with INSTALLS += qmldir, but with
CONFIG += builtin_resources (set for static) that's not the case.
AUX_QML_FILES is the correct variable though.

Task-number: QTBUG-67644
Change-Id: I6cf580586dec0bf7b44c89b95134d7d7f683102e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>