Create the decoration component when it is needed
If multiple QQmlEngines are created and use a Dialog,
initalizeEngine will be called more than once and therefore recreates
the QQuickAbstractDialog::m_decorationComponent without deleting the
original. Therefore we keep the decoration component url instead, and
create the decoration component on demand. This ensures that
it keeps the component relevant for the QQmlEngine and stops the
memory leak from before as well.

Fixes: QTBUG-75149
Change-Id: Ic79103f42d092d39b46868a45b4099ddc6edf9db
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>