Exclude the resources when building statically as qmake will handle this
As qmake will combine all the QML files etc together for us in static
mode then we should not include those added to the resources explicitly
in the pro file otherwise we will end up with an extra copy.
Additionally, if qrc:/ was added to the import path list it would end
trying to load the files from there and not from the
qrc:/qt-project.org/imports location as a result.

Change-Id: Icbaea7c6103074f9468914365b099be4589bee37
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>