Only calculate the height of the ListView when there are model entries
When it tries to calculate the height of the listview then it depends
on the model having entries. If there is no model or no entries in it
then it will cause a warning to be outputted indicating that the
model size is less than 0. This prevents it from occurring as a result.

Fixes: QTBUG-72543
Change-Id: I7155a3a240fdcdf93caddfce332f0cc6bb0f0ec0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>