Doc: Fix documentation warnings
All warnings in qtquickcontrols were related to \qmlmethod or \qmlsignal
parameters, except for the missing dependency to qtcore in Qt Quick
Extras.

There's some inconsistencies with how QDoc parses QML method parameters,
especially for qml signals - in some cases the \a command is replaced
with \e to clear the warning, even though the parameter name referenced
with \a does exist.

This commit brings the current warning count to zero.

Fixes: QTBUG-79828
Change-Id: I3d3d02583498c5a579f6476273e10e75680d940c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>