Update commercial license headers
Updated header.COMM to the files in tqtc-qtquickcontrols. Examples,
tests or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals
with the earlier license header. Reason for this
is that some autotests use hard coded line numbers
and a change in the line count causes failures in
tests.

Task-number: QTQAINFRA-4215
Change-Id: Id7a83ea4b83801352f9d45eaa688777add6784ab
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>