QQuickAbstractDialog: treat closing as rejection
QDialog treats the closing of the dialog via the [X] button as
rejection. QQuickAbstractDialog just recorded the change in its
m_visible member, but failed to emit rejected(), which caused users of
the dialog to wait indefinitely for a result.

Fix by emitting rejected() when the user closes the dialog via [X].

Fixes: QTBUG-101973
Change-Id: I5a66a69aec33f6a05279f20880e70cf633102a8d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>